+++ Banter/Core/Domain/MessageRepository.swift
@@ -7,10 +7,6 @@
 //
 
 protocol MessageRepository {
-    ///В качестве ``id`` поставил String
-    ///Более гибко получается без привязки к конкретным объектам
-    ///На выходе сделал ``Message`` но я думаю что тут надо будет сделать
-    ///протоколом, что бы не было привязки к конкретным объектам
-    func getMessages(id: /*Chat.ID*/ String) async throws -> [Message]
+    func getMessages(id: Chat.ID) async throws -> [Message]
   // messagesStream(id: Chat.ID) async throws -> AsyncSequence<[ChatMessage]>
 }
