+++ Banter/Scenes/Chat/Model/Message.swift
@@ -1,0 +1,12 @@
+//error nil
+
+
+import Foundation
+import MessageKit
+
+struct Message: MessageType{
+    var sender: SenderType
+    var messageId: String
+    var sentDate: Date
+    var kind: MessageKind
+}
