+++ Banter/Scenes/Chat/Model/Sender.swift
